pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-adapt"
version = "0.1.0"
description = "Observe. Adapt. Contribute. — CLI for the Adaptation Lifecycle"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Tobias Oberrauch" }]
keywords = ["cli", "adaptation", "upstream", "downstream", "github"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
code-adapt = "code_adapt.cli.main:app"
cadp = "code_adapt.cli.main:app"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0"]

[tool.hatch.build.targets.wheel]
packages = ["code_adapt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100
